added shared object error in TODO list
This error has been reported by rpmlint. It actually doesn't follow the naming
conventions. The only problem is that the shared object is built by make and
the version is owned by the rpm spec file. One solution can be to pass a make
variable containing %{version} throught command line.