added qs_version function and --version option
When querysort is built from the sources, a default value of "dev" is
defined for QS_VERSION. It is needed to define the version at comile
time with the -D QS_VERSION=\"x.y.z\" flag. If querysort is built using
make, a default value of 0.0.0 is set. The actual version is intended to
be set by the package builder (as for today, RPM only).