migration to autotools
QuerySort is no longer built with just `make'. In order to make it
portable, the following tools are used:
* autogen
* autoconf
* autoheader
* automake
* libtool

The shared object version doesn't follow the package version anymore,
this should guarantee  binary compatibility for future versions, but
also breack compatibility with 1.0.0.

The 1.0.0 is missing uriparser's license, which is now present. For this
reason I won't distribute querysort-1.0.0 binaries anymore.