major refactoring of the implementation
This implementation is inspired by object-oriented programming. A
query_sort struct contains fields and "abstract methods"
(implementation-dependant functions). Static functions can be seen as
concrete methods (and the first parameter as `this').

It makes future implementations easier to integrate by making the output
abstract : no matter how the result is written, the algorithm is always
the same.

The `qs_sort_clean' implementation is also more straightforward. It's
the same `qs_sort' algorithm with a slight variation indicated with a
flag. It doesn't need to write empty parameters and then move back
before those parameters.

It's all possible thanks to the incremental concatenations of the
result. The drawback of this approach is that a crash in the "abstract
methods" can lead to an inconsistent state where the sorted URL is
partially written.