man pages are created from RST templates
Since new features recently appeared, and others are expected, the
manual needed a little update. Switching to reStructuredText makes it
really easier to write without bringing too much boilerplate to the
build.

Unfortunately, the version number is hardcoded in the templates, which
is at least better than before since the version was initially missing.